fixed issue with hue rotation, still doesn't look right

+++ src/rtd/utils/input_image.py
@@ -317,15 +317,12 @@ class InputImageProcessor:
 
         # HSV color space adjustment
         if self.saturation != 1.0 or self.hue_rotation_angle != 0.0 or self.brightness != 1.0:
-            # Convert RGB to HSV (since our input is in RGB format)
+            # Convert to HSV
             img_hsv = cv2.cvtColor(img, cv2.COLOR_RGB2HSV)
             
-            # Convert to float for calculations
-            img_hsv = img_hsv.astype(np.float32)
-            
-            # Apply hue rotation (OpenCV uses 0-180 range for hue)
+            # Apply hue rotation
             if self.hue_rotation_angle != 0.0:
-                img_hsv[:, :, 0] = (img_hsv[:, :, 0] + self.hue_rotation_angle) % 180.0
+                img_hsv[:, :, 0] = (img_hsv[:, :, 0] + self.hue_rotation_angle) % 180
             
             # Apply saturation
             if self.saturation != 1.0:
@@ -335,8 +332,8 @@ class InputImageProcessor:
             if self.brightness != 1.0:
                 img_hsv[:, :, 2] = np.clip(img_hsv[:, :, 2] * self.brightness, 0, 255)
             
-            # Convert back to RGB (not BGR)
-            img = cv2.cvtColor(img_hsv.astype(np.uint8), cv2.COLOR_HSV2RGB)
+            # Convert back to BGR
+            img = cv2.cvtColor(img_hsv, cv2.COLOR_HSV2RGB)
         else:
             img = img.astype(np.uint8)
